backend/requirements.txt
fastapi
uvicorn
pyannote.audio
transformers
torch
Pillow
PyMuPDF
requests
huggingface_hub
